debug on vritualbox, mp is running!!!

+++ kernel/apic.h
@@ -10,14 +10,14 @@
 #define APIC_ICR_L	0x300
 #define APIC_ICR_H	0x310
 
-static inline u64 apic_read(u64 reg)
+static inline u32 apic_read(u64 reg)
 {
-        return *((volatile u64 *)(APIC_BASE_ADDR + reg));
+        return *((volatile u32 *)(APIC_BASE_ADDR + reg));
 }
 
-static inline void apic_write(u64 reg, u64 val)
+static inline void apic_write(u32 reg, u32 val)
 {
-        *((volatile u64 *)(APIC_BASE_ADDR + reg)) = val;
+        *((volatile u32 *)(APIC_BASE_ADDR + reg)) = val;
 }
 
 #endif
